+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
-- BREAKING CHANGE: Removed `yt-dlp` compile-time feature and made it a runtime switch (`use_yt_dlp()`)
-- feat: Implement downloading for both `yt-dlp` and `youtube-dl`
+- BREAKING CHANGE: Removed support for youtube-dl. Now only supports `yt-dlp`
+- feat: Add helper to download `yt-dlp` programatically with `reqwest`.
+- ci: set up GitHub Actions
 
 # 0.7.0
 - Added async support via `tokio`, disabled per default. You can opt-in via the `tokio` feature.